accountsからappにユーザ情報モデル移動
