+++ project/accounts/urls.py
@@ -6,7 +6,6 @@ from django.contrib.auth import views as auth_views
 app_name='accounts'
 
 urlpatterns=[
-    path('',views.user_page, name="top"),
     path('login/', auth_views.login, name='login'),
     path('logout/', auth_views.logout, name='logout'),
 ]
