+++ project/app/forms.py
@@ -1,6 +1,5 @@
 from django import forms
-from .models import Message
-from social_django.models import UserSocialAuth
+from .models import Message,Introduce
 
 # class MessageForm(forms.Form):
 #     message = forms.CharField(label='message', max_length=10000)
@@ -9,3 +8,9 @@ class MessageForm(forms.ModelForm):
     class Meta:
         model = Message # どのモデルからformつくる？
         fields = ('message',) # どのフィールドこのformで使う
+
+class IntroduceForm(forms.ModelForm):
+
+    class Meta:
+        model = Introduce
+        fields = ("company_name","recruiter",)
