+++ project/app/urls.py
@@ -6,7 +6,7 @@ app_name = 'app'
 urlpatterns = [
     path('', views.DashboardPage.as_view(), name='dashboard'),
     # path('user/', views.userProfilePage.as_view(), name='user'),
-    path('user/', include('accounts.urls'), name='user'),
+    path('user/', views.user_page, name='user'),
     path('tables/', views.tablesPage.as_view(), name='tables'),
     path('message/create/', views.get_message, name='create'),
     path('message/edit/<int:pk>', views.UpdateMessage.as_view(), name='edit'),
