+++ project/templates/accounts/top.html
@@ -1,14 +1,8 @@
 {% extends 'base.html' %}
 {% block content %}
-
 <div>
   <div>
-    <p><b>ScreenName</b>:{{ user.access_token.screen_name }}</p>
-    <p><b>UserId</b>:{{ user.access_token.user_id }}</p>
-    <p><b>OAuthTokenSecret</b>:{{ user.access_token.oauth_token_secret }}</p>
-    <p><b>OAuthToken</b>:{{ user.access_token.oauth_token }}</p>
-    <p><b>会社名</b>:{{ mess.company_name }}</p>
-    <p><b>採用者</b>:{{ mess.recruiter }}</p>
+    <p><b>Twitter Account Name</b>:{{ user.access_token.screen_name }}</p>
   </div>
 </div>
 {% with form as f %}
